Added deleting files.

+++ matebot-core/src/main/java/io/arusland/bots/commands/CommonCommand.java
@@ -30,18 +30,18 @@ public class CommonCommand extends BaseBotCommand {
         Message msg = update.getMessage();
 
         if (msg.hasText()) {
-            handleTextMessage(update, msg, user, storage, storage);
+            handleTextMessage(update, msg, user, storage);
         } else {
             sendMessage(msg.getChatId(), "It's not a text!");
         }
     }
 
-    private void handleTextMessage(Update update, Message message, User user, UserStorage userStorage, UserStorage storage) {
+    private void handleTextMessage(Update update, Message message, User user, UserStorage storage) {
         if (message.getText().startsWith(BotCommand.COMMAND_INIT_CHARACTER)) {
             ShortcutCommand cmd = getContext().getShortcutCommand(user, message.getText());
 
             if (cmd != null) {
-                handleShortcutCommand(cmd, user, update, storage);
+                handleShortcutCommand(cmd, user, update.getMessage().getChatId(), update, storage);
                 return;
             }
         }
@@ -63,10 +63,9 @@ public class CommonCommand extends BaseBotCommand {
         }
     }
 
-    private void handleShortcutCommand(ShortcutCommand cmd, User user, Update update, UserStorage storage) {
+    private void handleShortcutCommand(ShortcutCommand cmd, User user, Long chatId, Update update, UserStorage storage) {
         if (cmd.hasArguments()) {
             if ("cd".equals(cmd.getCommand())) {
-
                 getContext().setCurrentDir(user, cmd.getArguments().get(0));
                 getContext().listCurrentDir(update);
 
@@ -77,21 +76,42 @@ public class CommonCommand extends BaseBotCommand {
                     File file = item.tryGetFile();
 
                     if (file != null && !file.isDirectory()) {
-                        getContext().sendFile(update.getMessage().getChatId(), file);
+                        getContext().sendFile(chatId, file);
                     }
                 } else {
-                    sendMessage(update.getMessage().getChatId(), "Cannot find item: " + cmd.getArguments());
+                    sendMessage(chatId, "Cannot find item: " + cmd.getArguments());
                 }
             } else if ("rm".equals(cmd.getCommand())) {
                 Item item = storage.getItemByPath(cmd.getArguments().get(0));
 
                 if (item != null && !item.isDirectory()) {
-                    storage.deleteItem(item);
-                    getContext().listCurrentDir(update);
+                    if (cmd.getArguments().size() >= 2 && cmd.getArguments().get(1).equals("1")) {
+                        storage.deleteItem(item);
+                        sendMessage(chatId, "File '" + item.getFullPath() + "' removed");
+                        getContext().clearShortcutCommands(user);
+                        getContext().listCurrentDir(update);
+                    } else {
+                        String removeFile = "/remove";
+                        String cancelOperation = "/cancel";
+
+                        getContext().clearShortcutCommands(user);
+                        getContext().addShortcutCommand(user, removeFile, "rm", item.getFullPath(), "1");
+                        getContext().addShortcutCommand(user, cancelOperation, "cd", item.getParentPath());
+
+                        StringBuilder sb = new StringBuilder("Are you sure to remove file '");
+                        sb.append(item.getFullPath());
+                        sb.append("'?\n❌");
+                        sb.append(removeFile);
+                        sb.append(" ✅");
+                        sb.append(cancelOperation);
+                        sb.append("\n");
+
+                        sendMessage(chatId, sb.toString());
+                    }
                 }
             } else if ("mv".equals(cmd.getCommand())) {
                 // move file
-                sendMessage(update.getMessage().getChatId(), "TODO: Moving file: " + cmd.getArguments());
+                sendMessage(chatId, "TODO: Moving file: " + cmd.getArguments());
             } else {
                 System.out.println("Unknown command: " + cmd.getCommand() + " by shortcut: " + cmd.getShortcut());
             }
